fixed some problems reported by VC
git-svn-id: svn://172.20.112.140/eiddevel/trunk/projects/eIDClientCore/src@87 f977464c-999d-4b73-b5b9-8482498a25cc

+++ lib/epa-api/ePACommon.cpp
@@ -444,7 +444,7 @@ bool verifyResponse_AES(
 	// Store the SSC as vector.
 	std::vector<unsigned char> ssc_;
 	
-	for (int i = 0; i < kMac.size() - 8; i++)
+	for (size_t i = 0; i < kMac.size() - 8; i++)
 		ssc_.push_back(0x00);
 	
 	ssc_.push_back((ssc << 56) & 0xFF);
@@ -501,7 +501,7 @@ bool verifyResponse_AES(
 		vssc.push_back(0x00);
 	
 	std::vector<unsigned char> kMac_;
-	for (int i = 0; i < kMac.size(); i++)
+	for (size_t i = 0; i < kMac.size(); i++)
 		kMac_.push_back(kMac[i]);
 	
 	std::vector<unsigned char> calculatedMAC_ = calculateMAC(vssc, kMac_);
